% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_index.R
\name{forage_index}
\alias{forage_index}
\title{Calculate landscape forage quality index}
\usage{
forage_index(
  output_dir,
  landcover_path,
  foragetable_path = NA,
  forage_table,
  seasons,
  forage_range = NA,
  guild_table = NA,
  agg_factor = NA,
  normalize = T,
  useW = F,
  check_foragetable = T,
  rastertag = NA,
  compress_rasters = T
)
}
\arguments{
\item{output_dir}{Path to directory for output files.}

\item{landcover_path}{Path to land cover raster, including base file name}

\item{foragetable_path}{Path to forage quality by land use table}

\item{forage_table}{Forage quality by land use table (insead of foragetable_path)}

\item{seasons}{seasons to include. Must match names of forage table.}

\item{forage_range}{Foraging range (in m) to use for distance weighting forage scores surrounding focal cell.}

\item{guild_table}{Bee community to use to model foraging activity. Includes foraging range and relative abundnace of each species.

#optional parameters}

\item{agg_factor}{Aggregation factor for large rasters (use 4 to convert 50m to 120m resolution CDL)}

\item{normalize}{Normalize values by the number of cells within each moving window?}

\item{rastertag}{Text string to include in name of output raster}
}
\description{
New implemenatation of landsacpe forage quality index (Lonsdorf et al 2009). Uses FFT convolution to improve runtime.
}
\examples{


}
\keyword{forage}
\keyword{index}
