% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesting_index.R
\name{nesting_index}
\alias{nesting_index}
\title{Calculate landscape nesting quality index}
\usage{
nesting_index(
  output_dir,
  landcover_path,
  habitattable_path = NA,
  habitat_table,
  nest_locations,
  forage_range = NA,
  guild_table = NA,
  agg_factor = NA,
  normalize = T,
  useW = F,
  rastertag = NA,
  verbose = T
)
}
\arguments{
\item{output_dir}{Path to directory for output files.}

\item{landcover_path}{Path to land cover raster, including base file name}

\item{habitattable_path}{Path to habitat quality by land use table (CDL class integers should be in column called 'LULC')}

\item{habitat_table}{Habitat quality by land use table (instead of habitattable_path)}

\item{nest_locations}{Nesting types to include. Must match names of habitat table.}

\item{forage_range}{Foraging range (in m) to use for distance weighting scores surrounding focal cell.}

\item{guild_table}{Bee community to use to model foraging activity. Includes foraging range and relative abundance of each species.

#optional parameters}

\item{agg_factor}{Aggregation factor for large rasters (use 4 to convert 50m to 120m resolution CDL)}

\item{normalize}{Normalize values by the number of cells within each moving window?}

\item{rastertag}{Text string to include in name of output raster}

\item{verbose}{Include more log messages from model run?}
}
\description{
Variation of landscape nesting quality index (Lonsdorf et al 2009). Uses FFT convolution to improve runtime.
}
\details{
It is necessary to specify 'forage_range' OR 'guild_table,' not both.
}
\keyword{index}
\keyword{nesting}
